refactor: split catalog/queries.ts into focused modules
The catalog queries file was 1,260 lines with 37 exported functions
mixing WHERE builders, CTE builders, JOIN builders, CASE expressions,
and main query composers with no clear separation. This made it
extremely hard to navigate or understand the query structure.

Split into 4 focused files under queries/:
- filters.ts (~270 lines): all WHERE clause builders
- fragments.ts (~360 lines): CTEs, JOINs, CASE expressions, sort/pagination
- search.ts (~150 lines): search-specific queries (tag, name, utility)
- index.ts (~530 lines): main query composers + re-exports

The old queries.ts becomes a single re-export line so all external
imports continue working without changes.

No function implementations were modified — only code was moved
between files with updated import paths.
